DOSDoom 0.2, reconstructed
This isn't a guarantee of a recovered source, however it is hopefully
close.  It is derived from PDoom, which itself was lightly modified
from DOSDoom 0.2.  The effort was initiated by Linguica, and further
tweaked by Chungy and Quasar.

The time stamp is also not representative of the original source, but
we do have a release announcement confirming 1997-12-27.