Boom 2.02, DSL version (reconstructed)
This is a reconstruction effort by chungy, similar to how Linguica
reconstructed DOSDOOM 0.20.  It was done by replacing file headers
with older versions, with evidence from PrBoom 2.02 (see the prboom
branch).

The time stamp is also not represtantive of the original source.
However, we have a good approximation because of known release
announcements.